+++ lib/src/pdf_api.dart
@@ -5,11 +5,25 @@ import 'dart:ui' as ui;
 import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:http/http.dart' as http;
 
 // The trick to support Flutter Web is to use conditional import
 // Both of the files define PdfDocumentFactoryImpl class but only one of them is imported.
 import '../pdfrx.dart';
 import 'pdfium/pdfrx_pdfium.dart' if (dart.library.js) 'web/pdfrx_web.dart';
+
+/// Class to provide Pdfrx's configuration.
+abstract class Pdfrx {
+  /// Explicitly specify pdfium module path for special purpose.
+  ///
+  /// It is not supported on Flutter Web.
+  static String? pdfiumModulePath;
+
+  /// Overriding the default HTTP client for PDF download.
+  ///
+  /// It is not supported on Flutter Web.
+  static http.Client Function()? createHttpClient;
+}
 
 /// For platform abstraction purpose; use [PdfDocument] instead.
 abstract class PdfDocumentFactory {
@@ -270,12 +284,18 @@ abstract class PdfPage {
   int get pageNumber;
 
   /// PDF page width in points (width in pixels at 72 dpi) (rotated).
+  ///
+  /// To get width considering the rotation, use [getSize].
   double get width;
 
   /// PDF page height in points (height in pixels at 72 dpi) (rotated).
+  ///
+  /// To get width considering the rotation, use [getSize].
   double get height;
 
   /// PDF page size in points (size in pixels at 72 dpi) (rotated).
+  ///
+  /// To get width considering the rotation, use [getSize].
   Size get size => Size(width, height);
 
   /// PDF page rotation.
@@ -288,7 +308,8 @@ abstract class PdfPage {
   /// - If [x], [y] are not specified, (0,0) is used.
   /// - If [width], [height] is not specified, [fullWidth], [fullHeight] is used.
   /// - If [fullWidth], [fullHeight] are not specified, [PdfPage.width] and [PdfPage.height] are used (it means rendered at 72-dpi).
-  /// [backgroundColor] is used to fill the background of the page. If no color is specified, [Colors.white] is used.
+  /// - [backgroundColor] is used to fill the background of the page. If no color is specified, [Colors.white] is used.
+  /// - [rotationOverride] is used to override the page rotation. If not specified, [PdfPage.rotation] is used.
   /// - [annotationRenderingMode] controls to render annotations or not. The default is [PdfAnnotationRenderingMode.annotationAndForms].
   /// - [cancellationToken] can be used to cancel the rendering process. It must be created by [createCancellationToken].
   ///
@@ -315,6 +336,7 @@ abstract class PdfPage {
     double? fullWidth,
     double? fullHeight,
     Color? backgroundColor,
+    PdfPageRotation? rotationOverride,
     PdfAnnotationRenderingMode annotationRenderingMode =
         PdfAnnotationRenderingMode.annotationAndForms,
     PdfPageRenderCancellationToken? cancellationToken,
@@ -331,11 +353,24 @@ abstract class PdfPage {
   /// if [compact] is true, it tries to reduce memory usage by compacting the link data.
   /// See [PdfLink.compact] for more info.
   Future<List<PdfLink>> loadLinks({bool compact = false});
+
+  /// Get size.
+  ///
+  /// The function returns the size identical to [PdfPage.size] unless [rotationOverride] is specified.
+  ///
+  /// If [rotationOverride] is specified, the function returns the size considering the rotation.
+  Size getSize({PdfPageRotation? rotationOverride}) {
+    if (rotationOverride == null ||
+        ((rotationOverride.index - rotation.index) & 1) == 0) {
+      return size;
+    }
+    return Size(height, width);
+  }
 }
 
 /// Page rotation.
 enum PdfPageRotation {
-  none,
+  clockwise0,
   clockwise90,
   clockwise180,
   clockwise270,
@@ -756,6 +791,96 @@ class PdfTextRangeWithFragments {
   }
 }
 
+class PdfPageCoordsConverter {
+  PdfPageCoordsConverter(this.page,
+      {required this.pageRect, PdfPageRotation? rotationOverride})
+      : rotationOverride = rotationOverride ?? page.rotation;
+
+  PdfPageCoordsConverter withPageRect(Rect pageRect) =>
+      PdfPageCoordsConverter(page,
+          pageRect: pageRect, rotationOverride: rotationOverride);
+
+  final PdfPage page;
+  final Rect pageRect;
+  final PdfPageRotation rotationOverride;
+  late final bool isWidthHeightSwapped =
+      ((page.rotation.index + 4 - rotationOverride.index) & 1) == 1;
+  late final width = isWidthHeightSwapped ? page.height : page.width;
+  late final height = isWidthHeightSwapped ? page.width : page.height;
+  late final scale = pageRect.height / height;
+
+  /// Create a [PdfPoint] from X and Y coordinates.
+  PdfPoint fromOffset(double x, double y) {
+    x = x * width / pageRect.width;
+    y = y * height / pageRect.height;
+    switch (rotationOverride) {
+      case PdfPageRotation.clockwise0:
+        return PdfPoint(x, height - y);
+      case PdfPageRotation.clockwise90:
+        return PdfPoint(y, x);
+      case PdfPageRotation.clockwise180:
+        return PdfPoint(width - x, y);
+      case PdfPageRotation.clockwise270:
+        return PdfPoint(height - y, width - x);
+    }
+  }
+
+  /// Convert to [Rect] in Flutter coordinate.
+  Rect toRect(PdfRect rect) {
+    switch (rotationOverride) {
+      case PdfPageRotation.clockwise0:
+        final scale = pageRect.height / height;
+        return Rect.fromLTRB(
+          rect.left * scale,
+          (height - rect.top) * scale,
+          rect.right * scale,
+          (height - rect.bottom) * scale,
+        );
+      case PdfPageRotation.clockwise90:
+        final scale = pageRect.height / height;
+        return Rect.fromLTRB(
+          rect.top * scale,
+          rect.right * scale,
+          rect.bottom * scale,
+          rect.left * scale,
+        );
+      case PdfPageRotation.clockwise180:
+        final scale = pageRect.height / height;
+        return Rect.fromLTRB(
+          pageRect.width - rect.right * scale,
+          rect.bottom * scale,
+          pageRect.width - rect.left * scale,
+          rect.top * scale,
+        );
+      case PdfPageRotation.clockwise270:
+        final scale = pageRect.height / height;
+        return Rect.fromLTRB(
+          (width - rect.top) * scale,
+          pageRect.height - rect.left * scale,
+          (width - rect.bottom) * scale,
+          pageRect.height - rect.right * scale,
+        );
+    }
+  }
+
+  ///  Convert to [Rect] in Flutter coordinate using [pageRect] as the page's bounding rectangle.
+  Rect toRectWithPageOffset(PdfRect rect) =>
+      toRect(rect).translate(pageRect.left, pageRect.top);
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+    return other is PdfPageCoordsConverter &&
+        other.page == page &&
+        other.pageRect == pageRect &&
+        other.rotationOverride == rotationOverride;
+  }
+
+  @override
+  int get hashCode =>
+      page.hashCode ^ pageRect.hashCode ^ rotationOverride.hashCode;
+}
+
 /// Rectangle in PDF page coordinates.
 ///
 /// Please note that PDF page coordinates is different from Flutter's coordinate.
@@ -810,67 +935,6 @@ class PdfRect {
   /// Empty rectangle.
   static const empty = PdfRect(0, 0, 0, 0);
 
-  /// Convert to [Rect] in Flutter coordinate.
-  /// [page] is the page to convert the rectangle.
-  /// [scaledPageSize] is the scaled page size to scale the rectangle. If not specified, [PdfPage.size] is used.
-  /// [rotation] is the rotation of the page. If not specified, [PdfPage.rotation] is used.
-  Rect toRect({
-    required PdfPage page,
-    Size? scaledPageSize,
-    int? rotation,
-  }) {
-    final rotated = rotate(rotation ?? page.rotation.index, page);
-    final scale =
-        scaledPageSize == null ? 1.0 : scaledPageSize.height / page.height;
-    return Rect.fromLTRB(
-      rotated.left * scale,
-      (page.height - rotated.top) * scale,
-      rotated.right * scale,
-      (page.height - rotated.bottom) * scale,
-    );
-  }
-
-  ///  Convert to [Rect] in Flutter coordinate using [pageRect] as the page's bounding rectangle.
-  Rect toRectInPageRect({
-    required PdfPage page,
-    required Rect pageRect,
-  }) =>
-      toRect(page: page, scaledPageSize: pageRect.size)
-          .translate(pageRect.left, pageRect.top);
-
-  PdfRect rotate(int rotation, PdfPage page) {
-    final swap = (page.rotation.index & 1) == 1;
-    final width = swap ? page.height : page.width;
-    final height = swap ? page.width : page.height;
-    switch (rotation & 3) {
-      case 0:
-        return this;
-      case 1:
-        return PdfRect(
-          bottom,
-          width - left,
-          top,
-          width - right,
-        );
-      case 2:
-        return PdfRect(
-          width - right,
-          height - bottom,
-          width - left,
-          height - top,
-        );
-      case 3:
-        return PdfRect(
-          height - top,
-          right,
-          height - bottom,
-          left,
-        );
-      default:
-        throw ArgumentError.value(rotate, 'rotate');
-    }
-  }
-
   PdfRect inflate(double dx, double dy) =>
       PdfRect(left - dx, top + dy, right + dx, bottom - dy);
 
@@ -923,6 +987,58 @@ extension PdfRectsExt on Iterable<PdfRect> {
     }
     return PdfRect(left, top, right, bottom);
   }
+}
+
+/// Pdf point in PDF page coordinates.
+@immutable
+class PdfPoint {
+  const PdfPoint(this.x, this.y);
+
+  /// Create a [PdfPoint] from X and Y coordinates.
+  /// [page] is the page to convert the rectangle.
+  /// [scaledPageSize] is the scaled page size to scale the point. If not specified, no scaling is applied.
+  /// [rotationOverride] is the rotation of the page. If not specified, [PdfPage.rotation] is used.
+  factory PdfPoint.fromXy(
+    double x,
+    double y, {
+    required PdfPage page,
+    Size? scaledPageSize,
+    PdfPageRotation? rotationOverride,
+  }) {
+    if (scaledPageSize != null) {
+      x = x * page.width / scaledPageSize.width;
+      y = y * page.height / scaledPageSize.height;
+    }
+    rotationOverride ??= page.rotation;
+    switch (rotationOverride.index & 3) {
+      case 0:
+        return PdfPoint(x, page.height - y);
+      case 1:
+        return PdfPoint(y, x);
+      case 2:
+        return PdfPoint(page.width - x, y);
+      case 3:
+        return PdfPoint(page.height - y, page.width - x);
+      default:
+        throw ArgumentError.value(rotationOverride, 'rotation');
+    }
+  }
+
+  /// X coordinate in PDF page.
+  final double x;
+
+  /// Y coordinate in PDF page.
+  final double y;
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+
+    return other is PdfPoint && other.x == x && other.y == y;
+  }
+
+  @override
+  int get hashCode => x.hashCode ^ y.hashCode;
 }
 
 /// PDF [Explicit Destination](https://opensource.adobe.com/dc-acrobat-sdk-docs/pdfstandards/PDF32000_2008.pdf#page=374) the page and inner-page location to jump to.